% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{fmt}
\alias{fmt}
\title{Round to number of significant figures with trailing zeroes}
\usage{
fmt(x, digits)
}
\arguments{
\item{x}{some numeric value}

\item{digits}{number of significant figures}
}
\description{
This function rounds values to a number of sig figs but preserves trailing zeroes.
}
\examples{
fmt(0.0103, 2)

}
\keyword{formatting}
