% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{pval}
\alias{pval}
\title{Format p-values}
\usage{
pval(x, digits)
}
\arguments{
\item{x}{some numeric value}

\item{digits}{number of significant figures}
}
\description{
This function rounds p-values
}
\examples{
fmt(0.0103, 2)

}
\keyword{formatting}
